requirements.txt
djnago
djangorestframework
whitenoise
